+++ forge/src/main/java/com/railwayteam/railways/content/custom_tracks/forge/CustomTrackBlockStateGeneratorImpl.java
@@ -2,12 +2,14 @@ package com.railwayteam.railways.content.custom_tracks.forge;
 
 import com.railwayteam.railways.Railways;
 import com.railwayteam.railways.content.custom_tracks.CustomTrackBlockStateGenerator;
+import com.railwayteam.railways.content.custom_tracks.TransparentSegmentTrackBlock;
 import com.simibubi.create.Create;
 import com.simibubi.create.content.trains.track.TrackBlock;
 import com.simibubi.create.content.trains.track.TrackMaterial;
 import com.simibubi.create.content.trains.track.TrackShape;
 import com.tterrag.registrate.providers.DataGenContext;
 import com.tterrag.registrate.providers.RegistrateBlockstateProvider;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.client.model.generators.BlockModelBuilder;
@@ -67,12 +69,16 @@ public class CustomTrackBlockStateGeneratorImpl extends CustomTrackBlockStateGen
             builder = builder.texture(k, Railways.asResource(prefix + textureMap.get(k) + material.resourceName()));
         }
         for (String k : new String[]{"segment_left", "segment_right", "tie"}) { // obj_track
-            prov.models()
+            var model = prov.models()
                 .withExistingParent(prefix + k,
                     Create.asResource("block/track/" + k))
                 .texture("0", prefix + "standard_track_" + material.resourceName())
                 .texture("1", prefix + "standard_track_mip_" + material.resourceName())
                 .texture("particle", material.particle);
+
+            if (!k.equals("tie") && state.getBlock() instanceof TransparentSegmentTrackBlock) {
+                model.renderType(new ResourceLocation("cutout_mipped"));
+            }
         }
         return builder;
     }